added message edit re-trigger

+++ src/events/closeableModules/AutoResponder.ts
@@ -6,6 +6,7 @@ import {CloseOptionModel} from "../../model/DB/autoMod/impl/CloseOption.model";
 import {TriggerConstraint} from "../../model/closeableModules/impl/TriggerConstraint";
 import {AutoResponderModel} from "../../model/DB/autoMod/impl/AutoResponder.model";
 import {Message} from "discord.js";
+import {MessageEventEditTrigger} from "../../model/decorators/MessageEventEditTrigger";
 
 export class AutoResponder extends TriggerConstraint {
 
@@ -15,9 +16,10 @@ export class AutoResponder extends TriggerConstraint {
         super(CloseOptionModel, AutoResponder._uid);
     }
 
+    @MessageEventEditTrigger
     @On("message")
     @Guard(NotBot, EnabledGuard("AutoResponder"))
-    private async process([message]: ArgsOf<"message">, client: Client): Promise<void> {
+    private async process([message]: ArgsOf<"message">, client: Client, isUpdate = false): Promise<void> {
         const guildId = message.guild.id;
         const channel = message.channel;
         const allRespondObjects = await AutoResponderModel.findAll({
@@ -56,6 +58,9 @@ export class AutoResponder extends TriggerConstraint {
             const {responseType} = autoResponder;
             switch (responseType) {
                 case "message": {
+                    if (isUpdate) {
+                        continue;
+                    }
                     const responseText: string = await this._parseVars(autoResponder.response, message);
                     if (publicDelete) {
                         message.delete();
